sdm845-common: implement dynamic thermal profile with Pixel thermal-engine
Parts:
Instead of writing to sconfig in sysfs, change vendor.thermal.vr_mode
property so that the init will restart thermal-engine service to apply
the new config file. This mechanism was originally for VR/non-VR
profile switch, but it can also be utilized to apply other profiles.

Only gaming profile is kept because the default profile is more than
enough for other use cases. A separate config file for things like
camera and streaming is not needed for now. The code is still flexible
enough to accommodate additional profiles in the future.

configs:
A copy of "-novr" thermal engine config file is made for "-gaming".
Changes to that config file will be made in a later commit.

sepolicy:
Move vendor_thermal_prop property type declaration to public so that
both XiaomiParts and thermal HAL can access it.

Test: set "gaming" profile for an app, launch that app, and check if
      `getprop vendor.thermal.vr_mode` returns "-gaming". Then return
      to HOME screen and check if the prop is set back to "-novr"
Test: check logcat to see if thermal-engine is restarted with the
      correct config file

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Signed-off-by: Dhruv <dhruvgera61@gmail.com>